pyproject.toml
[tool.poetry]
name = "rag-race"
version = "0.1.0"
description = ""
authors = ["bhanpuramufaddal <bhanpuramufaddal@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_demo"}]

[tool.poetry.dependencies]
python = "^3.12"
tqdm = "^4.66.4"
datasets = "^2.20.0"
sentencepiece = "^0.2.0"
qdrant-client = "^1.10.1"
ipykernel = "^6.29.5"
torch = {version = "2.2", extras = ["cpu"]}
sentence-transformers = "^3.0.1"
numpy = "1.26.4"
dspy-ai = "^2.4.12"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
